requirements.txt
python-dotenv
psycopg
ruff
django-stubs
